+++ features/test/jjs105-devcontainer/install-lib-test.sh
@@ -1,0 +1,12 @@
+#-------------------------------------------------------------------------------
+# Copyright (c) Jon Spain (https://github.com/jjs105).
+# Licensed under GNU GPL v3 or later
+# https://github.com/jjs105/devcontainers/tree/main/LICENSE
+#-------------------------------------------------------------------------------
+
+# This file is not intended to be run as a script, although it is possible.
+# Rather it should be sourced from the main development container feature test
+# script.
+
+# Test script for the development container install  library contained in the
+# jjs105-devcontainer development container feature.